requirements.txt
pandas
requests
python-dotenv
sqlalchemy
